Struct merkle::ProofBlock

pub struct ProofBlock {
    pub node_hash: Vec<u8>,
    pub sibling_hash: Positioned<Vec<u8>>,
    pub sub_proof: Option<Box<ProofBlock>>,
}

A ProofBlock is a linked-list holding the hash of the node, the hash of its sibling node, and the rest of the inclusion proof.

Fields
node_hash: Vec<u8> sibling_hash: Positioned<Vec<u8>> sub_proof: Option<Box<ProofBlock>>
Methods

impl ProofBlock

fn new<T>(tree: &Tree<T>, needle: &Vec<u8>) -> Option<ProofBlock> where T: Hashable

Attempt to generate a proof that the hash needle is a member of the given tree.