Struct merkle::MerkleTree

pub struct MerkleTree<D, T> {
    // some fields omitted
}

A Merkle tree is a binary tree, with values of type T at the leafs, and where every node holds the hash of the concatenation of the hashes of its children nodes.

Methods

impl<D, T> MerkleTree<D, T> where D: Digest + Clone, T: Into<Vec<u8>> + Clone

fn from_vec_unsafe(digest: D, values: Vec<T>) -> Self

Constructs a Merkle Tree from a vector of data blocks. WARNING: Panics if values is empty!

fn from_vec(digest: D, values: Vec<T>) -> Option<Self>

Constructs a Merkle Tree from a vector of data blocks. Returns None if values is empty.

fn digest(&self) -> &D

Returns the hash function used in this Merkle tree

fn root_hash(&self) -> &Vec<u8>

Returns the root hash of Merkle tree

fn height(&self) -> usize

Returns the height of Merkle tree

fn count(&self) -> usize

Returns the number of leaves in the Merkle tree

fn gen_proof(&self, value: &T) -> Option<Proof<D, T>>

Generate an inclusion proof for the given value. Returns None if the given value is not found in the tree.