Struct merkle::MerkleTree

pub struct MerkleTree<T> {
    pub algorithm: &'static Algorithm,
    // some fields omitted
}

A Merkle tree is a binary tree, with values of type T at the leafs, and where every internal node holds the hash of the concatenation of the hashes of its children nodes.

Fields
algorithm: &'static Algorithm

The hashing algorithm used by this Merkle tree

Methods

impl<T> MerkleTree<T> where T: Into<Vec<u8>> + Clone

fn from_vec(algorithm: &'static Algorithm, values: Vec<T>) -> Option<Self>

Constructs a Merkle Tree from a vector of data blocks. Returns None if values is empty.

fn root_hash(&self) -> &Vec<u8>

Returns the root hash of Merkle tree

fn height(&self) -> usize

Returns the height of Merkle tree

fn count(&self) -> usize

Returns the number of leaves in the Merkle tree

fn gen_proof(&self, value: T) -> Option<Proof<T>>

Generate an inclusion proof for the given value. Returns None if the given value is not found in the tree.

fn iter(&self) -> LeavesIterator<T>

Creates an Iterator over the values contained in this Merkle tree.